Changed TME_INDEX_DictIter to TME_INDEX_KeysIter Added !v->ob_type check
In JyNI_GetJythonDelegate(in JyNI.c), there is now a check in case an
ob_type hasn't been initialised with PyType_Ready()
This could go in PyObject_Call, but if it is in JyNI_GetJythonDelegate
then it may get called in other places and prevent other bugs of this
type. But it means the if statement is called more frequently which
could slow things down. A discussion could be had about the best place.